Relocating automowerAccessoryFactory class.
