+++ src/automowerAccessoryFactory.ts
@@ -1,26 +1,26 @@
 import { API, PlatformAccessory } from 'homebridge';
 
-import { AutomowerAccessory, AutomowerContext } from '../automowerAccessory';
-import { PlatformLogger } from '../diagnostics/platformLogger';
-import { Mower } from '../model';
-import { AccessoryInformation, AccessoryInformationImpl } from '../services/accessoryInformation';
-import { ArrivingContactSensorImpl, ArrivingSensor } from '../services/arrivingSensor';
-import { MowerControlServiceImpl } from '../services/husqvarna/automower/mowerControlService';
-import { BatteryInformation, BatteryInformationImpl } from '../services/batteryInformation';
-import { LeavingContactSensorImpl, LeavingSensor } from '../services/leavingSensor';
-import { MotionSensor, MotionSensorImpl } from '../services/motionSensor';
-import { PauseSwitch, PauseSwitchImpl } from '../services/pauseSwitch';
-import { DeterministicMowerFaultedPolicy } from '../services/policies/mowerFaultedPolicy';
-import { DeterministicMowerInMotionPolicy } from '../services/policies/mowerInMotionPolicy';
-import { DeterministicMowerIsArrivingPolicy } from '../services/policies/mowerIsArrivingPolicy';
-import { DeterministicMowerIsLeavingPolicy } from '../services/policies/mowerIsLeavingPolicy';
-import { DeterministicMowerIsPausedPolicy } from '../services/policies/mowerIsPausedPolicy';
-import { DeterministicMowerTamperedPolicy } from '../services/policies/mowerTamperedPolicy';
-import { DeterministicScheduleEnabledPolicy } from '../services/policies/scheduleEnabledPolicy';
-import { ScheduleSwitch, ScheduleSwitchImpl } from '../services/scheduleSwitch';
-import { Localization } from './localization';
-import { PlatformAccessoryFactory } from './platformAccessoryFactory';
-import { PlatformContainer } from './platformContainer';
+import { AutomowerAccessory, AutomowerContext } from './automowerAccessory';
+import { PlatformLogger } from './diagnostics/platformLogger';
+import { Mower } from './model';
+import { AccessoryInformation, AccessoryInformationImpl } from './services/accessoryInformation';
+import { ArrivingContactSensorImpl, ArrivingSensor } from './services/arrivingSensor';
+import { MowerControlServiceImpl } from './services/husqvarna/automower/mowerControlService';
+import { BatteryInformation, BatteryInformationImpl } from './services/batteryInformation';
+import { LeavingContactSensorImpl, LeavingSensor } from './services/leavingSensor';
+import { MotionSensor, MotionSensorImpl } from './services/motionSensor';
+import { PauseSwitch, PauseSwitchImpl } from './services/pauseSwitch';
+import { DeterministicMowerFaultedPolicy } from './services/policies/mowerFaultedPolicy';
+import { DeterministicMowerInMotionPolicy } from './services/policies/mowerInMotionPolicy';
+import { DeterministicMowerIsArrivingPolicy } from './services/policies/mowerIsArrivingPolicy';
+import { DeterministicMowerIsLeavingPolicy } from './services/policies/mowerIsLeavingPolicy';
+import { DeterministicMowerIsPausedPolicy } from './services/policies/mowerIsPausedPolicy';
+import { DeterministicMowerTamperedPolicy } from './services/policies/mowerTamperedPolicy';
+import { DeterministicScheduleEnabledPolicy } from './services/policies/scheduleEnabledPolicy';
+import { ScheduleSwitch, ScheduleSwitchImpl } from './services/scheduleSwitch';
+import { Localization } from './primitives/localization';
+import { PlatformAccessoryFactory } from './primitives/platformAccessoryFactory';
+import { PlatformContainer } from './primitives/platformContainer';
 
 /**
  * A mechanism to create {@link AutomowerAccessory} instances.
